Created Demo data type
