+++ server/src/main/java/com/feeder/server/model/GenericData.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 public abstract class GenericData {
 
   public enum Type {
+    DEMO,
     REDDIT,
     GITHUB,
     SPOTIFY,
