+++ server/src/main/java/com/feeder/server/model/SpotifyData.java
@@ -1,7 +1,3 @@
 package com.feeder.server.model;
 
-import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
-
-@JsonInclude(Include.NON_EMPTY)
 public abstract class SpotifyData extends GenericData {}
